Renaming from disabled to unsavedChanges
